+++ workspaces/boost/plugins/boost-common/src/index.ts
@@ -29,3 +29,4 @@ export const BOOST_PLUGIN_ID = 'boost';
 
 export * from './types';
 export * from './permissions';
+export * from './ingestion-health';